requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.111.0
APScheduler==3.9.1
httpx==0.27.0
pytrends==4.9.2
requests==2.31.0
